streamlit/requirements.txt
altair
pandas
streamlit
pymongo
plotly-express
streamlit-autorefresh
ipywidgets>=7.0.0